Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

I'm using this on Supersonic and Vision. It completely unborked the GPS on Supersonic and improves the lock speed on Vision immensely.

Tested it on Stacie's MT4G, but she's not had any issues with GPS so far.

Patch-set: 1